+++ R/auxiliary_ecdf.R
@@ -2,7 +2,8 @@
 # ecdf_check     : check whether an input is a valid list of ecdfs
 # ecdf_quantiles : return the processed data (grid size=1000)
 # ecdf_wsum      : compute a weighted sum
-# ecdf_2dist     : distance between two discretized quantiles
+# ecdf_2dist     : 2-distance between two discretized quantiles
+# ecdf_pdist     : p-distance between two discretized quantiles
 
 # ecdf_check --------------------------------------------------------------
 #' @keywords internal
@@ -64,4 +65,15 @@ ecdf_2dist <- function(xvec, yvec1, yvec2){
   ydsq = ((yvec1-yvec2)^2)
   n    = length(ydsq)
   return(sqrt(sum(((ydsq[1:(n-1)]+ydsq[2:n])/2)*base::diff(xvec))))
+}
+
+
+# ecdf_pdist --------------------------------------------------------------
+#' @keywords internal
+#' @noRd
+ecdf_pdist <- function(xvec, yvec1, yvec2, p){
+  ydsq   = abs(yvec1-yvec2)^p
+  n      = length(ydsq)
+  output = sum(((ydsq[1:(n-1)]+ydsq[2:n])/2)*base::diff(xvec))
+  return(output^(1/p))
 }